tests folder and move preeip155, 1559 tests into
